refactor(fundraiser): modularize GoFundMe data synchronization logic
Extract GoFundMe fetching and file persistence logic into dedicated
utility functions and a standalone script to improve maintainability
and allow for manual updates.

- move data fetching and saving logic to `utils/fundraiser.ts`
- simplify Netlify function to use the new utility
- add `update-fundraiser.ts` script for manual execution via `tsx`
- add `update-fundraiser` npm script
- remove redundant `types/fundraiser.ts` in favor of utility-local types
- allow `public/fundraiser.json` to be tracked in version control
