+++ package.json
@@ -13,7 +13,8 @@
     "lint": "oxlint",
     "lint:fix": "oxlint --fix",
     "lint:github": "oxlint --format=github",
-    "icons": "nuxt-seo-utils icons --source favicon.svg"
+    "icons": "nuxt-seo-utils icons --source favicon.svg",
+    "update-fundraiser": "tsx update-fundraiser.ts"
   },
   "dependencies": {
     "@netlify/functions": "^5.3.0",
@@ -35,9 +36,11 @@
     "@nuxt/icon": "^1.12.0",
     "@nuxtjs/tailwindcss": "^6.13.0",
     "@tailwindcss/typography": "^0.5.16",
+    "@types/node": "^25.9.1",
     "nuxt-seo-utils": "^8.1.11",
     "oxfmt": "^0.50.0",
     "oxlint": "^1.65.0",
+    "tsx": "^4.22.4",
     "typescript": "^5.7.0"
   },
   "packageManager": "pnpm@11.0.9"
