+++ public/fundraiser.json
@@ -1,0 +1,1 @@
+{"currentAmount":{"amount":36886,"currencyCode":"EUR"},"goalAmount":{"amount":40000,"currencyCode":"EUR"},"donationCount":925,"donationsEnabled":true}
